+++ src/components/HappinessForm.vue
@@ -23,10 +23,10 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "vue-property-decorator";
+import { Component, Vue } from "vue-property-decorator";
 
 @Component
-export default class HappinessChart extends Vue {}
+export default class HappinessForm extends Vue {}
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
